+++ backend/src/main/java/com/MaFederation/MaFederation/repository/StaffRepository.java
@@ -2,12 +2,18 @@ package com.MaFederation.MaFederation.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.MaFederation.MaFederation.dto.Staff.ResponceStaffDTO;
 import com.MaFederation.MaFederation.model.Staff;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
 
-public interface StaffRepository extends JpaRepository<Staff, Long> {}
+public interface StaffRepository extends JpaRepository<Staff, Integer> 
+
+{
+   
+
+}
